rbutilqt/themeeditor: Fix compiler warnings
- match sign in integer comparison
- remove unused RBViewport member

+++ utils/rbutilqt/base/ttscarbon.cpp
@@ -328,7 +328,7 @@ int TTSCarbon::convertAiffToWav(const char* aiff, const char* wav)
     if(!buf) /* error out if malloc() failed */
         return -1;
     in = fopen(aiff, "rb");
-    if(fread(buf, 1, filestat.st_size, in) < filestat.st_size) {
+    if(fread(buf, 1, filestat.st_size, in) < (size_t)filestat.st_size) {
         printf("could not read file: not enought bytes read\n");
         fclose(in);
         free(buf);
